+++ src/features/translation/ui/translation-container.tsx
@@ -126,24 +126,77 @@ export const TranslationContainer = () => {
     setIsAutoCopy1(checked);
   };
 
+  const [isCopied1, setIsCopied1] = useState(false);
+  const [isCopied2, setIsCopied2] = useState(false);
+
+  const handleCopy = async (
+    ref: React.RefObject<HTMLTextAreaElement>,
+    setCopied: (value: boolean) => void
+  ) => {
+    if (ref.current) {
+      await navigator.clipboard.writeText(ref.current.value);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    }
+  };
+
+  const [hasMounted, setHasMounted] = useState(false);
+
+  useEffect(() => {
+    setHasMounted(true);
+  }, []);
+
+  useEffect(() => {
+    const handleKeyDown = (e: globalThis.KeyboardEvent) => {
+      const isMac = /Mac|iPod|iPhone|iPad/.test(navigator.userAgent);
+      const modifierKey = isMac ? e.metaKey : e.ctrlKey;
+      if (modifierKey) {
+        handleLanguageShortcut(e.key as TranslationModeKeysType);
+        if (TranslationMode[e.key as TranslationModeKeysType]) {
+          e.preventDefault();
+        }
+      }
+    };
+
+    document.addEventListener("keydown", handleKeyDown);
+    return () => document.removeEventListener("keydown", handleKeyDown);
+  }, [handleLanguageShortcut]);
+
+  const inputRef = useRef<HTMLTextAreaElement | null>(null);
+
+  useEffect(() => {
+    if (inputRef.current) {
+      inputRef.current.focus();
+    }
+  }, [hasMounted, inputRef]);
+
+  if (!hasMounted) {
+    return null;
+  }
+
   return (
     <Translation
       register={register}
       errors={errors}
-      input={inputValue}
-      selectedModel={selectedModelValue}
+      input={watch("input")}
+      selectedModel={watch("selectedModel")}
       isLoading={isSubmitting}
       models={MODELS}
       onKeyDown={handleKeyDown}
       onPaste={handlePaste}
-      onSubmit={submitForm}
+      onSubmit={handleSubmit(onTranslationSubmit)}
       translation1={translation1}
       translation2={translation2}
       translation1Ref={translation1RefObject}
       translation2Ref={translation2RefObject}
-      onLanguageShortcut={handleLanguageShortcut}
       onAutoCopyChange1={handleAutoCopyChange1}
       isAutoCopy1={isAutoCopy1}
+      isCopied1={isCopied1}
+      setIsCopied1={setIsCopied1}
+      isCopied2={isCopied2}
+      setIsCopied2={setIsCopied2}
+      handleCopy={handleCopy}
+      inputRef={inputRef}
     />
   );
 };
